+++ tests/dem/particle_particle_contact_force_linear.cc
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2020-2025 The Lethe Authors
+// SPDX-FileCopyrightText: Copyright (c) 2020-2026 The Lethe Authors
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
 
 /**
@@ -25,26 +25,23 @@ test()
   triangulation.refine_global(refinement_number);
   MappingQ<dim> mapping(1);
 
-
   // Defining general simulation parameters
   DEMSolverParameters<dim>                              dem_parameters;
-  Parameters::Lagrangian::LagrangianPhysicalProperties &lagrangian_prop =
+  Parameters::Lagrangian::LagrangianPhysicalProperties &lpp =
     dem_parameters.lagrangian_physical_properties;
-  Parameters::Lagrangian::ModelParameters<dim> &model_param =
-    dem_parameters.model_parameters;
 
   set_default_dem_parameters(1, dem_parameters);
 
   Tensor<1, dim> g{{0, 0, -9.81}};
-  double         dt                                        = 0.00001;
-  double         particle_diameter                         = 0.005;
-  lagrangian_prop.particle_type_number                     = 1;
-  lagrangian_prop.youngs_modulus_particle[0]               = 50000000;
-  lagrangian_prop.poisson_ratio_particle[0]                = 0.3;
-  lagrangian_prop.restitution_coefficient_particle[0]      = 0.5;
-  lagrangian_prop.friction_coefficient_particle[0]         = 0.5;
-  lagrangian_prop.rolling_friction_coefficient_particle[0] = 0.1;
-  lagrangian_prop.density_particle[0]                      = 2500;
+  double         dt                            = 0.00001;
+  double         particle_diameter             = 0.005;
+  lpp.particle_type_number                     = 1;
+  lpp.youngs_modulus_particle[0]               = 50000000;
+  lpp.poisson_ratio_particle[0]                = 0.3;
+  lpp.restitution_coefficient_particle[0]      = 0.5;
+  lpp.friction_coefficient_particle[0]         = 0.5;
+  lpp.rolling_friction_coefficient_particle[0] = 0.1;
+  lpp.density_particle[0]                      = 2500;
 
   const double neighborhood_threshold = std::pow(1.3 * particle_diameter, 2);