+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the Lethe project will be documented in this file.
 The changelog for the previous releases of Lethe are located in the release_notes folder.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Master] - 2026/02/09
+
+### Added
+
+- MINOR This PR adds the option to declare a ``custom distribution`` as ``volume`` or ``number`` weighted. In addition, an option was added to the "custom distribution" allowing the sampled diameter values to be interpolated between input data points. Finally, the ``custom distribution`` diameter values and probability values can be define using the probability density function (PDF) or the cummulative density function (CDF) of the particle size distribution (PSD). [1892](https://github.com/chaos-polymtl/lethe/pull/1892)
+
+### Changed 
+
+- MINOR Containers for the lagrangian particle properties are changed from ``std::unordered_map`` to ``std::vectors``.  [1892](https://github.com/chaos-polymtl/lethe/pull/1892)
+
 ## [Master] - 2026/02/07
 
 ### Changed